+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Mzinga Changelog #
 
+## next ##
+
+* Viewer: Fixed issue with loading newer BoardSpace SGF files which record pass moves differently
+
 ## v0.13.4 ##
 
 * Viewer: Engine Console now includes engine's std:err output